Add support for SafeByte and SafeBytes
Currently, there is no direct way to write a safe byte or a byte slice
to a SafeWriter. They have to be type-casted to either a rune (for byte)
or a string (for []byte). It would be more efficient to add
the types (and helpers) SafeByte and SafeBytes.

Also, UnsafeByte and UnsafeBytes already exist. So, introducing SafeByte and
SafeBytes would also make the API more consistent.